fix typechecking

+++ mteb/_evaluators/classification_metrics.py
@@ -8,7 +8,7 @@ if TYPE_CHECKING:
     from numpy.typing import NDArray
 
 
-def hamming_score(y_true: NDArray[np.int64], y_pred: NDArray[np.int64]) -> float:
+def hamming_score(y_true: NDArray[np.integer], y_pred: NDArray[np.integer]) -> float:
     """Compute the Hamming score (a.k.a. label-based accuracy) for multilabel classification.
 
     The Hamming score is the fraction of labels that are correctly predicted for each sample,
